# Update JSON platform benchmarks to use JSON source generator