--- a/src/BenchmarksApps/Kestrel/PlatformBenchmarks/BenchmarkApplication.Caching.cs
+++ b/src/BenchmarksApps/Kestrel/PlatformBenchmarks/BenchmarkApplication.Caching.cs
@@ -10,7 +10,11 @@ namespace PlatformBenchmarks
     {
         private async Task Caching(PipeWriter pipeWriter, int count)
         {
+#if NET6_0_OR_GREATER
             OutputMultipleQueries(pipeWriter, await RawDb.LoadCachedQueries(count), SerializerContext.CachedWorldArray);
+#else
+            OutputMultipleQueries(pipeWriter, await RawDb.LoadCachedQueries(count));
+#endif
         }
     }
 }