| Profile | Scenario | RPS (Before) | RPS (After) | RPS % increase | Max alloc rate (Before) (b/sec) | Max alloc rate (After) (b/sec) | Max alloc rate % decrease | Latency (Before) (ms) | Latency (After) (ms) | Latency % decrease | Start time + first request (Before) (ms) | Start time (After) | % Start time decrease |
| --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
| aspnet-citrine-lin | json | 1,232,440 | 1,265,521 | 2.684187466 | 41,672,184 | 12,429,712 | 70.17264082 | 0.57 | 0.51 | 10.52631579 | 246 | 262 | -6.504065041 |
| | multiple\_queries | 23,452 | 23,434 | -0.07675251578 | 752,562,891 | 727,026,011 | 3.393321715 | 10.92 | 11.01 | -0.8241758242 | 1725 | 1735 | -0.5797101449 |
| | single\_query | 397,556 | 402,294 | 1.191781787 | 1,084,997,824 | 1,098,367,805 | -1.232258785 | 0.66 | 0.66 | 0 | 1940 | 1831 | 5.618556701 |
| | updates | 18,276 | 18,478 | 1.105274677 | 1,072,226,685 | 1,067,972,931 | 0.3967215198 | 14 | 13.86 | 1 | 1955 | 1961 | -0.3069053708 |

We see a small increase in RPS and a big drop in max alloc rate for the `json` benchmark. There's also a small increase in RPS for the `single_query` benchmark. Otherwise, the rest of the benchmarks don't change substantially.

Benchmarks were run in 6 iterations; excluding the worst two and the best result; then taking the mean.
Benchmarks had 30s warmup.
Benchmarks run on `aspnet-citrine-lin` profile.

```
crank --config platform.benchmarks.yml --scenario json --profile aspnet-citrine-lin --variable warmup=30 --iterations 6 --exclude 2:1 --exclude-order "load:wrk/rps/mean;http/rps/mean" --application.options.collectCounters true
```


FYI @sebastienros @adamsitnik @eerhardt @steveharter @ericstj @stephentoub @jkotas @davidfowl