@adamsitnik 2021-06-21
@layomia I was looking forward to find out what is the difference for Caching benchmark, where so far 40% of time was spent for JSON serialization. So I run the benchmarks against your fork:

| load                   |    net5.0 |    net6.0 |   this PR |
| ---------------------- | --------- | --------- | --------- |
| CPU Usage (%)          | 20        | 21        | 30        |
| Cores usage (%)        | 556       | 597       | 842       |
| Working Set (MB)       | 37        | 38        | 37        |
| Private Memory (MB)    | 358       | 358       | 358       |
| Start Time (ms)        | 0         | 0         | 0         |
| First Request (ms)     | 66        | 67        | 62        |
| **Requests/sec**           | **243,000**   | **260,928**   | **364,224**   |
| Requests               | 3,669,151 | 3,939,804 | 5,499,468 |
| Mean latency (ms)      | 1.22      | 1.09      | 0.90      |
| Max latency (ms)       | 57.32     | 47.40     | 44.63     |
| Bad responses          | 0         | 0         | 0         |
| Socket errors          | 0         | 0         | 0         |
| Read throughput (MB/s) | 764.86    | 821.29    | 1,146.88  |
| Latency 50th (ms)      | 1.03      | 0.96      | 0.67      |
| Latency 75th (ms)      | 1.14      | 1.06      | 0.76      |
| Latency 90th (ms)      | 1.28      | 1.19      | 0.87      |
| Latency 99th (ms)      | 9.57      | 6.34      | 9.40      |

That gives us a **100k RPS GAIN** which is almost `+30%`.

cc @jeffhandley @maryamariyan