@layomia 2021-07-19
@sebastienros the 3.1 and 5.0 implementations are unchanged in this PR, meaning that they don't use the source generator and things shouldn't break. This is based on ifdef'ing the new code based on `NET6_0_OR_GREATER`. FWIW the generator works well with these TFMs, but we're deciding not to use it for those cases.